Bumped to version 23
This version surfaces a GetExtensionFromMimeType function, a
GenerateFileName function, and custom media downloader interface.

+++ Disa.Framework/Bubbles/AudioBubble.cs
@@ -71,13 +71,13 @@ namespace Disa.Framework.Bubbles
 		[NonSerialized]
 		public bool ControlPause = true;
 		[NonSerialized]
-		public int ControlPosition;
+		public long ControlPosition;
 		[NonSerialized]
 		public Action ControlPositionUpdated;
 		[NonSerialized]
 		public Action ControlStopped;
 		[NonSerialized]
-		public int ControlDuration;
+		public long ControlDuration;
 
 
         public AudioBubble(long time, BubbleDirection direction, string address,
